requirements.txt
Django==5.2.1
gunicorn==23.0.0
whitenoise==6.9.0
dj-database-url==2.3.0
python-dotenv==1.1.0
psycopg==3.2.7
psycopg-binary==3.2.7
pillow==11.2.1
djangorestframework==3.16.0
django-import-export==4.3.7

